moved some of the functioanlity from the locate can task to the navigation file
