--- Rover_Control/include/main.h
@@ -1,10 +1,0 @@
-#ifndef MAIN_H
-#define MAIN_H
-
-#define LEFT 1
-#define RIGHT 0
-
-void setInitialOffset(void);
-void setOffsetBasedOnOneSide(bool);
-
-#endif
